fix: textshape angle in ISO camera
